requirements.txt
matplotlib==2.2.2
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.10.2
django-celery==3.1.17
django-datatables-view==1.14.0
django-debug-toolbar==1.5
django-extensions==1.7.4
django-filter==0.15.3
django-push-notifications==1.2.1
djangorestframework==3.5.3
docopt==0.6.2
docutils==0.12
fcm-django==0.2.3
gunicorn==19.3.0
httplib2==0.11.3
Jinja2==2.8
kombu==3.0.37
Markdown==2.6.1
MarkupSafe==0.23
numpy==1.14.5
oauth2==1.9.0.post1
oauthlib==2.0.0
Pillow==2.8.1
psycopg2==2.7.3
pydocs==0.1
pyfcm==1.0.1
python-dateutil==2.5.3
python-magic==0.4.12
pytz==2015.2
redis==2.10.3
requests==2.11.1
requests-oauthlib==0.7.0
reverse-geocoder==1.5.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.10.0
sklearn==0.0
sqlparse==0.2.1
twython==3.3.0

